+++ lib/app/view/app.dart
@@ -1,6 +1,6 @@
 import 'package:qruto_budget/accounts_list/repository/account_repository.dart';
 import 'package:qruto_budget/accounts_list/repository/account_repository_drift.dart';
-import 'package:qruto_budget/backup/cubit/auth_cubit.dart';
+import 'package:qruto_budget/backup/cubit/backup_cubit.dart';
 import 'package:qruto_budget/categories/repository/category_repository.dart';
 import 'package:qruto_budget/categories/repository/category_repository_drift.dart';
 import 'package:qruto_budget/charts/repository/chart_repository.dart';
@@ -44,7 +44,7 @@ class App extends StatelessWidget {
             create: (context) => ThemeCubit(),
           ),
           BlocProvider(
-            create: (context) => AuthCubit()..checkUserStatus(),
+            create: (context) => BackupCubit()..checkUserStatus(),
           ),
         ],
         child: AppView(),