+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## [4.?.?] - ??-??-????
 ### Added
+- Google Drive backup
 - Lock app orientation to portrait mode
 ### Fixed
-
+- Font sizes
 
 ## [4.2.2] - 03-19-2025
 ### Added